implement lots of shit. stupid racecondition still exists :(

+++ utils/memutil.h
@@ -6,5 +6,5 @@
 
 extern "C" void* memcpy(void *dest, void const *src, size_t size);
 extern "C" void* memmove(void *dest, void const *src, size_t size);
-extern "C" void* memset(void *dest, int pattern, size_t size);
+extern "C" void* memset(void *dest, uint8_t pattern, size_t size);
